Creates index.html and style.css
This is the basic landing page from the project design. There are some
things that could be closer to the image provided by The Odin Project,
but all in all it seems close enough as a starting point.

The plan is to come back to this project when I have learned more and
personalize it, while keeping the same basic structure. By that point
I'll have a more solid grasp over flexbox and can spend more time
designing a page that will make it stand out.